KKJJ-24 <refactor> (<User>) : 폴드명 수정후 import오류 수정
import설정이 바뀌지 않아 재수정
