+++ src/main/java/renewal/awesome_travel/config/CustomOAuth2UserService.java
@@ -1,7 +1,6 @@
 package renewal.awesome_travel.config;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
@@ -9,11 +8,11 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import renewal.awesome_travel.config.security.CustomUserDetails;
-import renewal.awesome_travel.member.entity.User;
-import renewal.awesome_travel.member.repository.UserRepository;
 import renewal.awesome_travel.member.utils.Provider;
 import renewal.awesome_travel.member.utils.Role;
 import renewal.awesome_travel.member.utils.Status;
+import renewal.awesome_travel.user.entity.User;
+import renewal.awesome_travel.user.repository.UserRepository;
 
 import java.util.Map;
