+++ src/main/java/renewal/awesome_travel/config/security/CustomUserDetailsService.java
@@ -5,8 +5,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import renewal.awesome_travel.member.entity.User;
-import renewal.awesome_travel.member.repository.UserRepository;
+import renewal.awesome_travel.user.entity.User;
+import renewal.awesome_travel.user.repository.UserRepository;
 
 @Service
 @RequiredArgsConstructor
